Libkedin-AI-Agent/modules/publisher.py
import requests
import json
from config import LINKEDIN_ACCESS_TOKEN, LINKEDIN_MEMBER_ID


def publish_post(content: str):
    author_urn = f"urn:li:member:{LINKEDIN_MEMBER_ID}"

    headers = {
        "Authorization": f"Bearer {LINKEDIN_ACCESS_TOKEN}",
        "X-Restli-Protocol-Version": "2.0.0",
        "Content-Type": "application/json"
    }

    post_data = {
        "author": author_urn,
        "lifecycleState": "PUBLISHED",
        "specificContent": {
            "com.linkedin.ugc.ShareContent": {
                "shareCommentary": {
                    "text": content
                },
                "shareMediaCategory": "NONE"
            }
        },
        "visibility": {
            "com.linkedin.ugc.MemberNetworkVisibility": "PUBLIC"
        }
    }

    response = requests.post(
        "https://api.linkedin.com/v2/ugcPosts",
        headers=headers,
        data=json.dumps(post_data)
    )

    return response.status_code, response.text